+++ coloraide/average.py
@@ -2,13 +2,27 @@
 from __future__ import annotations
 import math
 from . import util
+import itertools as it
 from .spaces import HWBish
 from .types import ColorInput, AnyColor
 from typing import Iterable
 
+
+class Sentinel(float):
+    """Sentinel object that is specific to averaging that we shouldn't see defined anywhere else."""
+
+
+def _iter_colors(colors: Iterable[ColorInput]) -> Iterable[tuple[ColorInput, float]]:
+    """Iterate colors and return weights."""
+
+    for c in colors:
+        yield c, 1.0
+
+
 def average(
     color_cls: type[AnyColor],
     colors: Iterable[ColorInput],
+    weights: Iterable[float] | None,
     space: str,
     premultiplied: bool = True
 ) -> AnyColor:
@@ -18,6 +32,7 @@ def average(
     Polar coordinates use a circular mean: https://en.wikipedia.org/wiki/Circular_mean.
     """
 
+    sentinel = Sentinel()
     obj = color_cls(space, [])
 
     # Get channel information
@@ -31,15 +46,39 @@ def average(
     channels = cs.channels
     chan_count = len(channels)
     alpha_index = chan_count - 1
-    sums = [0.0] * chan_count
-    totals = [0.0] * chan_count
+    avgs = [0.0] * chan_count
+    counts = [0] * chan_count
     sin = 0.0
     cos = 0.0
+    wavg = 0.0
+    no_weights = weights is None
+    if no_weights:
+        weights = ()
+    mx = 0.0
 
-    # Sum channel values
-    i = -1
-    for c in colors:
-        obj.update(c)
+    # Sum channel values using a rolling average. Apply premultiplication and additional weighting as required.
+    count = 0
+    for c, w in (_iter_colors(colors) if no_weights else it.zip_longest(colors, weights, fillvalue=sentinel)):  # type: ignore[arg-type]
+
+        # Handle explicit weighted cases
+        if not no_weights:
+            # If there are more weights than colors, ignore additional weights
+            if c is sentinel:
+                break
+
+            # If there are less weights than colors, assume full weight for colors without weights
+            if w is sentinel:
+                w = mx
+
+            # Negative weights are considered as zero weight
+            if w < 0.0:
+                w = 0.0
+
+            # Track the largest weight so we can populate colors with no weights
+            elif w > mx:
+                mx = w
+
+        obj.update(c)  # type: ignore[arg-type]
         # If cylindrical color is achromatic, ensure hue is undefined
         if hue_index >= 0 and not math.isnan(obj[hue_index]) and obj.is_achromatic():
             obj[hue_index] = math.nan
@@ -47,51 +86,58 @@ def average(
         alpha = coords[-1]
         if math.isnan(alpha):
             alpha = 1.0
+        walpha = alpha * w
+        count += 1
+        wavg += (w - wavg) / count
         i = 0
         for coord in coords:
             # No need to average an undefined value or color components if alpha is zero
-            if not math.isnan(coord) and (premultiplied or alpha or i == alpha_index):
-                totals[i] += 1
+            is_alpha = i == alpha_index
+            if not math.isnan(coord) and (premultiplied or alpha or is_alpha):
+                counts[i] += 1
+                n = counts[i]
                 if i == hue_index:
                     rad = math.radians(coord)
                     if premultiplied:
-                        sin += math.sin(rad) * alpha
-                        cos += math.cos(rad) * alpha
+                        sin += ((math.sin(rad) * walpha) - sin) / n
+                        cos += ((math.cos(rad) * walpha) - cos) / n
                     else:
-                        sin += math.sin(rad)
-                        cos += math.cos(rad)
+                        sin += ((math.sin(rad) * w) - sin) / n
+                        cos += ((math.cos(rad) * w) - cos) / n
                 else:
-                    sums[i] += (coord * alpha) if premultiplied and i != alpha_index else coord
+                    avgs[i] += (((coord * walpha) if premultiplied and not is_alpha else (coord * w)) - avgs[i]) / n
             i += 1
 
-    if i == -1:
+    if not count:
         raise ValueError('At least one color must be provided in order to average colors')
 
-    # Get the mean
-    sums[-1] = alpha = math.nan if not totals[-1] else (sums[-1] / totals[-1])
+    # Undo premultiplication and weighting to get the final color
+    w_factor = math.nan if not wavg else wavg
+    avgs[-1] = alpha = math.nan if not counts[-1] else avgs[-1] / w_factor
     if math.isnan(alpha):
         alpha = 1.0
+    walpha = alpha * w_factor
+
     for i in range(chan_count - 1):
-        total = totals[i]
-        if not total or not alpha:
-            sums[i] = math.nan
+        if not counts[i] or not alpha:
+            avgs[i] = math.nan
         elif i == hue_index:
             if premultiplied:
-                sin /= total * alpha
-                cos /= total * alpha
+                sin /= walpha
+                cos /= walpha
             else:
-                sin /= total
-                cos /= total
+                sin /= w_factor
+                cos /= w_factor
             if abs(sin) < util.ACHROMATIC_THRESHOLD_SM and abs(cos) < util.ACHROMATIC_THRESHOLD_SM:
-                sums[i] = math.nan
+                avgs[i] = math.nan
             else:
                 avg_theta = math.degrees(math.atan2(sin, cos))
-                sums[i] = (avg_theta + 360) if avg_theta < 0 else avg_theta
+                avgs[i] = (avg_theta + 360) if avg_theta < 0 else avg_theta
         else:
-            sums[i] /= (total * alpha) if premultiplied else total
+            avgs[i] /= walpha if premultiplied else w_factor
 
-    # Create the color and if polar and there is no defined hue, force an achromatic state.
-    color = obj.update(space, sums[:-1], sums[-1])
+    # Create the color. If polar and there is no defined hue, force an achromatic state.
+    color = obj.update(space, avgs[:-1], avgs[-1])
     if cs.is_polar():
         if is_hwb and math.isnan(color[hue_index]):
             w, b = cs.indexes()[1:]  # type: ignore[attr-defined]