+++ tools/gamut_3d_plotly.py
@@ -713,6 +713,81 @@ def plot_interpolation(
     fig.add_trace(trace)
 
 
+def plot_average(
+    fig,
+    space,
+    avg_colors,
+    avg_space,
+    gmap,
+    simulate_alpha,
+    avg_gmap
+):
+    """Plot interpolations."""
+
+    if not avg_colors:
+        return
+
+    parts = avg_colors.split(':')
+
+    colors = parts[0].split(';')
+    if len(parts) == 2:
+        weights = [float(i.strip()) for i in parts[1].split(',')]
+    else:
+        weights = None
+
+    color = Color.average(
+        colors,
+        weights,
+        space=avg_space
+    )
+
+    target = Color.CS_MAP[space]
+    flags = {
+        'is_cyl': target.is_polar(),
+        'is_labish': isinstance(target, Labish),
+        'is_lchish': isinstance(target, LChish),
+        'is_hslish': isinstance(target, HSLish),
+        'is_hwbish': isinstance(target, HWBish),
+        'is_hsvish': isinstance(target, HSVish)
+    }
+
+    for s in colors:
+        x = []
+        y = []
+        z = []
+        cmap = []
+        c = Color(s).convert(space, in_place=True).normalize(nans=False)
+        if avg_gmap:
+            c.fit('srgb', **gmap)
+        store_coords(c, x, y, z, flags)
+        c.convert('srgb', in_place=True)
+        c.fit(**gmap)
+        if simulate_alpha:
+            cmap.append(Color.layer([c, 'white'], space='srgb').to_string(hex=True))
+        else:
+           cmap.append(c.to_string(comma=True, alpha=False))
+
+        c = Color(color).convert(space, in_place=True).normalize(nans=False)
+        if avg_gmap:
+            c.fit('srgb', **gmap)
+        store_coords(c, x, y, z, flags)
+        c.convert('srgb', in_place=True)
+        c.fit(**gmap)
+        if simulate_alpha:
+            cmap.append(Color.layer([c, 'white'], space='srgb').to_string(hex=True))
+        else:
+            cmap.append(c.to_string(comma=True, alpha=False))
+
+        trace = go.Scatter3d(
+            x=x, y=y, z=z,
+            marker={'size': [8, 16], 'color': cmap, 'opacity': 1},
+            line={'color': cmap[0], 'width': 3},
+            showlegend=False
+        )
+
+        fig.add_trace(trace)
+
+
 def main():
     """Main."""
 
@@ -756,14 +831,15 @@ def main():
     )
 
     # Interpolation visualization
+    parser.add_argument('--avg-colors', default='', help="Colors that should be averaged together.")
     parser.add_argument('--interp-colors', default='', help='Interpolation colors separated by semicolons.')
     parser.add_argument('--interp-method', default='linear', help="Interplation method to use: linear, bezier, etc.")
-    parser.add_argument('--interp-space', default='oklab', help="Interpolation space.")
+    parser.add_argument('--interp-space', default='oklab', help="Interpolation/averaging space.")
     parser.add_argument(
-        '--interp-alpha', action='store_true', help="Simulate interpolation opacity by overlaying on white"
+        '--interp-alpha', action='store_true', help="Simulate interpolation/averaging opacity by overlaying on white."
     )
     parser.add_argument(
-        '--interp-gmap', action='store_true', help="Force plotted interpolation results to be gamut mapped."
+        '--interp-gmap', action='store_true', help="Force plotted interpolation/averaging results to be gamut mapped."
     )
     parser.add_argument('--hue', default='shorter', help="Hue interpolation handling.")
     parser.add_argument('--extrapolate', action='store_true', help='Extrapolate values.')
@@ -856,6 +932,16 @@ def main():
         args.powerless,
         args.extrapolate,
         args.steps,
+        gmap,
+        args.interp_alpha,
+        args.interp_gmap
+    )
+
+    plot_average(
+        fig,
+        args.space,
+        args.avg_colors,
+        args.interp_space,
         gmap,
         args.interp_alpha,
         args.interp_gmap