+++ docs/src/markdown/about/changelog.md
@@ -3,6 +3,7 @@
 ## 4.7
 
 -   **NEW**: Officially support Python 3.14.
+-   **NEW**: `average()` now accepts weights for weighted averaging.
 -   **ENHANCE**: Switch to deploying with PyPI's "Trusted Publisher".
 -   **ENHANCE**: Performance improvements for various algebraic calculations.
 -   **FIX**: Fix some corner cases with some algebraic calculations.